+++ samplesheets/vueapp/tests/unit/IrodsDirModal.spec.js
@@ -135,6 +135,29 @@ describe('IrodsDirModal.vue', () => {
     expect(wrapper.findAll('.sodar-ss-irods-obj').at(1).isVisible()).toBe(true)
   })
 
+  it('clears filter input on modal re-open', async () => {
+    fetchMock.mock(listAjaxUrl, objList)
+    const wrapper = mount(IrodsDirModal, {
+      localVue, propsData: propsData
+    })
+    wrapper.vm.showModal(rootIrodsPath)
+    await waitNT(wrapper.vm)
+    await waitRAF()
+
+    expect(wrapper.find('#sodar-ss-irods-filter').element.value).toBe('')
+    await wrapper.find('#sodar-ss-irods-filter').setValue('test2')
+    expect(wrapper.find('#sodar-ss-irods-filter').element.value).toBe('test2')
+    expect(wrapper.find('.sodar-irods-obj-table').exists()).toBe(true)
+    wrapper.vm.hideModal()
+    await waitNT(wrapper.vm)
+    await waitRAF()
+
+    wrapper.vm.showModal(rootIrodsPath)
+    await waitNT(wrapper.vm)
+    await waitRAF()
+    expect(wrapper.find('#sodar-ss-irods-filter').element.value).toBe('')
+  })
+
   it('updates modal title on setTitle()', async () => {
     fetchMock.mock(listAjaxUrl, objList)
     const updatedTitle = 'Updated title'