+++ CHANGELOG.rst
@@ -46,6 +46,7 @@ Fixed
     - Crash from incompatibility with ``packaging==0.22`` (#1550)
     - Cancer shortcuts expecting specific naming convention (#1554)
     - Cancer shortcut caching with identical library names in study (#1560)
+    - iRODS modal filter input not cleared on modal re-open (#1555)
 
 Removed
 -------